+++ game/dota_addons/barebones/scripts/vscripts/libraries/notifications.lua
@@ -1,4 +1,4 @@
-NOTIFICATIONS_VERSION = "0.80"
+NOTIFICATIONS_VERSION = "0.85"
 
 --[[
   Sample Panorama Notifications Library by BMD
@@ -9,8 +9,8 @@ NOTIFICATIONS_VERSION = "0.80"
 
   Usage
   -Notifications can be sent to the Top or Bottom notification panel of an individual player, a whole team, or all clients at once.
-  -Notifications can be sent in pieces consisting of Labels and HeroImages.
-  -Both types of notifications have 4 potential parameters:
+  -Notifications can be sent in pieces consisting of Labels, Images, HeroImages, and AbilityImages.
+  -Notifications are specified by a table which has 4 potential parameters:
     -duration: The duration to display the notification for on screen.  Ignored for a notification which "continues" a previous notification line.
     -class: An optional (leave as nil for default) string which will be used as the class to add to the notification piece.
     -style: An optional (leave as nil for default) table of css properties to add to this notification, such as {["font-size"]="60px", color="green"}.
@@ -21,40 +21,38 @@ NOTIFICATIONS_VERSION = "0.80"
   -For HeroImages, there is two additional parameters:
     -hero:  (Mandatory) The hero name, e.g. "npc_dota_hero_axe".
     -imagestyle:  (Optional)  The image style to display for this hero image.  Default when 'nil' is 'icon'.  'portrait' and 'landscape' are two other options.
-  -Call the Notifications:Top, Notifications:TopToAll, or Notifications:TopToTeam to send a top-area notifications to the appropriate players 
-  -Call the Notifications:TopHeroImage, Notifications:TopHeroImageToAll, or Notifications:TopHeroImageToTeam to send a top-area heroimage notifications to the appropriate players 
-  -Call the Notifications:Bottom, Notifications:BottomToAll, or Notifications:BottomToTeam to send a bottom-area notifications to the appropriate players 
-  -Call the Notifications:BottomHeroImage, Notifications:BottomHeroImageToAll, or Notifications:BottomHeroImageToTeam to send a top-area heroimage notifications to the appropriate players 
+  -For AboilityImages, there is one additional mandatory parameter:
+    -ability:  The ability name, e.g. "lina_fiery_soul".
+  -For Images, there is one additional mandatory parameter:
+    -image:  The image src string, e.g. "file://{images}/status_icons/dota_generic.psd".
 
-  --Notifications can be given in function parameter form or as a table of values
+  -Call the Notifications:Top, Notifications:TopToAll, or Notifications:TopToTeam to send a top-area notification to the appropriate players 
+  -Call the Notifications:Bottom, Notifications:BottomToAll, or Notifications:BottomToTeam to send a bottom-area notifications to the appropriate players 
   
   Examples:
 
   -- Send a notification to all players that displays up top for 5 seconds
-  Notifications:TopToAll("Top Notification for 5 seconds", 5.0)
+  Notifications:TopToAll({text="Top Notification for 5 seconds ", duration=5.0})
+  -- Send a notification to playerID 0 which will display up top for 9 seconds and be green, on the same line as the previous notification
+  Notifications:Top(0, {text="GREEEENNNN", duration=9, style={color="green"}, continue=true})
+
+  -- Display 3 styles of hero icons on the same line for 5 seconds.
+  Notifications:TopToAll({hero="npc_dota_hero_axe", duration=5.0})
+  Notifications:TopToAll({hero="npc_dota_hero_axe", imagestyle="landscape", continue=true})
+  Notifications:TopToAll({hero="npc_dota_hero_axe", imagestyle="portrait", continue=true})
+
+  -- Display a generic image and then 2 ability icons on the same line for 5 seconds
+  Notifications:TopToAll({image="file://{images}/status_icons/dota_generic.psd", duration=5.0})
+  Notifications:TopToAll({ability="nyx_assassin_mana_burn", continue=true})
+  Notifications:TopToAll({ability="lina_fiery_soul", continue=true})
+
 
   -- Send a notification to all players on radiant (GOODGUYS) that displays near the bottom of the screen for 10 seconds to be displayed with the NotificationMessage class added
-  Notifications:BottomToTeam(DOTA_TEAM_GOODGUYS, "AAAAAAAAAAAAAA", 10, "NotificationMessage")
-
-  -- Table-based version of the above:
   Notifications:BottomToTeam(DOTA_TEAM_GOODGUYS, {text="AAAAAAAAAAAAAA", duration=10, class="NotificationMessage"})
-
-  -- Send a notification to playerID 0 which will display up top for 9 seconds and be green
-  Notifications:Top(0, "GREEEENNNN", 9, nil, {color="green"})
-
   -- Send a notification to player 0 which will display near the bottom a large red notification with a solid blue border for 5 seconds
-  Notifications:Bottom(PlayerResource:GetPlayer(0), "Super Size Red", 5, nil, {color="red", ["font-size"]="110px", border="10px solid blue"})
+  Notifications:Bottom(PlayerResource:GetPlayer(0), {text="Super Size Red", duration=5, style={color="red", ["font-size"]="110px", border="10px solid blue"}})
 
-  -- Table-based version of the above:
-  Notifications:Bottom(PlayerResource:GetPlayer(0), 
-    {text="Super Size Red", duration=5, style={color="red", ["font-size"]="110px", border="10px solid blue"}}
-  )
 
-  -- Send a 4-part notification consisting of a HeroImage icon, a small text piece, a large localized text piece, and a HeroImage portrait
-  Notifications:TopHeroImage(PlayerResource:GetPlayer(0), "npc_dota_hero_axe", nil, 30, nil, {width="64px", height="64px"}, false)
-  Notifications:Top(PlayerResource:GetPlayer(0), "1234 ASDF FAFF ", 3, nil, {["font-size"]="72px"}, true)
-  Notifications:Top(PlayerResource:GetPlayer(0), " asdf !@#$", 3, nil, {["font-size"]="42px", color="red"}, true)
-  Notifications:TopHeroImage(PlayerResource:GetPlayer(0), "npc_dota_hero_axe", "portrait", 30, nil, nil, true)
 
 ]]
 
@@ -62,120 +60,95 @@ if Notifications == nil then
   Notifications = class({})
 end
 
-function Notifications:Top(player, text, duration, class, style, continue)
+function Notifications:Top(player, table)
   if type(player) == "number" then
     player = PlayerResource:GetPlayer(player)
   end
 
-  if type(text) == "table" then
-    CustomGameEventManager:Send_ServerToPlayer(player, "top_notification", {text=text.text, duration=text.duration, class=text.class, style=text.style, continue=text.continue} )
+  if table.text ~= nil then
+    CustomGameEventManager:Send_ServerToPlayer(player, "top_notification", {text=table.text, duration=table.duration, class=table.class, style=table.style, continue=table.continue} )
+  elseif table.hero ~= nil then
+    CustomGameEventManager:Send_ServerToPlayer(player, "top_notification", {hero=table.hero, imagestyle=table.imagestyle, duration=table.duration, class=table.class, style=table.style, continue=table.continue} )
+  elseif table.image ~= nil then
+    CustomGameEventManager:Send_ServerToPlayer(player, "top_notification", {image=table.image, duration=table.duration, class=table.class, style=table.style, continue=table.continue} )
+  elseif table.ability ~= nil then
+    CustomGameEventManager:Send_ServerToPlayer(player, "top_notification", {ability=table.ability, duration=table.duration, class=table.class, style=table.style, continue=table.continue} )
   else
-    CustomGameEventManager:Send_ServerToPlayer(player, "top_notification", {text=text, duration=duration, class=class, style=style, continue=continue} )
+    CustomGameEventManager:Send_ServerToPlayer(player, "top_notification", {text="No TEXT provided.", duration=table.duration, class=table.class, style=table.style, continue=table.continue} )
   end
 end
 
-function Notifications:TopToAll(text, duration, class, style, continue)
-  if type(text) == "table" then
-    CustomGameEventManager:Send_ServerToAllClients( "top_notification", {text=text.text, duration=text.duration, class=text.class, style=text.style, continue=text.continue} )
+function Notifications:TopToAll(table)
+  if table.text ~= nil then
+    CustomGameEventManager:Send_ServerToAllClients("top_notification", {text=table.text, duration=table.duration, class=table.class, style=table.style, continue=table.continue} )
+  elseif table.hero ~= nil then
+    CustomGameEventManager:Send_ServerToAllClients("top_notification", {hero=table.hero, imagestyle=table.imagestyle, duration=table.duration, class=table.class, style=table.style, continue=table.continue} )
+  elseif table.image ~= nil then
+    CustomGameEventManager:Send_ServerToAllClients("top_notification", {image=table.image, duration=table.duration, class=table.class, style=table.style, continue=table.continue} )
+  elseif table.ability ~= nil then
+    CustomGameEventManager:Send_ServerToAllClients("top_notification", {ability=table.ability, duration=table.duration, class=table.class, style=table.style, continue=table.continue} )
   else
-    CustomGameEventManager:Send_ServerToAllClients( "top_notification", {text=text, duration=duration, class=class, style=style, continue=continue} )
+    CustomGameEventManager:Send_ServerToAllClients("top_notification", {text="No TEXT provided.", duration=table.duration, class=table.class, style=table.style, continue=table.continue} )
   end
 end
 
-function Notifications:TopToTeam(team, text, duration, class, style, continue)
-  if type(text) == "table" then
-    CustomGameEventManager:Send_ServerToTeam(team, "top_notification", {text=text.text, duration=text.duration, class=text.class, style=text.style, continue=text.continue} )
+function Notifications:TopToTeam(team, table)
+  if table.text ~= nil then
+    CustomGameEventManager:Send_ServerToTeam(team, "top_notification", {text=table.text, duration=table.duration, class=table.class, style=table.style, continue=table.continue} )
+  elseif table.hero ~= nil then
+    CustomGameEventManager:Send_ServerToTeam(team, "top_notification", {hero=table.hero, imagestyle=table.imagestyle, duration=table.duration, class=table.class, style=table.style, continue=table.continue} )
+  elseif table.image ~= nil then
+    CustomGameEventManager:Send_ServerToTeam(team, "top_notification", {image=table.image, duration=table.duration, class=table.class, style=table.style, continue=table.continue} )
+  elseif table.ability ~= nil then
+    CustomGameEventManager:Send_ServerToTeam(team, "top_notification", {ability=table.ability, duration=table.duration, class=table.class, style=table.style, continue=table.continue} )
   else
-    CustomGameEventManager:Send_ServerToTeam(team, "top_notification", {text=text, duration=duration, class=class, style=style, continue=continue} )
+    CustomGameEventManager:Send_ServerToTeam(team, "top_notification", {text="No TEXT provided.", duration=table.duration, class=table.class, style=table.style, continue=table.continue} )
   end
 end
 
 
-function Notifications:TopHeroImage(player, hero, imagestyle, duration, class, style, continue)
+function Notifications:Bottom(player, table)
   if type(player) == "number" then
     player = PlayerResource:GetPlayer(player)
   end
 
-  if type(hero) == "table" then
-    CustomGameEventManager:Send_ServerToPlayer(player, "top_notification_heroimage", {hero=hero.hero, imagestyle=hero.imagestyle, duration=hero.duration, class=hero.class, style=hero.style, continue=hero.continue} )
+  if table.text ~= nil then
+    CustomGameEventManager:Send_ServerToPlayer(player, "bottom_notification", {text=table.text, duration=table.duration, class=table.class, style=table.style, continue=table.continue} )
+  elseif table.hero ~= nil then
+    CustomGameEventManager:Send_ServerToPlayer(player, "bottom_notification", {hero=table.hero, imagestyle=table.imagestyle, duration=table.duration, class=table.class, style=table.style, continue=table.continue} )
+  elseif table.image ~= nil then
+    CustomGameEventManager:Send_ServerToPlayer(player, "bottom_notification", {image=table.image, duration=table.duration, class=table.class, style=table.style, continue=table.continue} )
+  elseif table.ability ~= nil then
+    CustomGameEventManager:Send_ServerToPlayer(player, "bottom_notification", {ability=table.ability, duration=table.duration, class=table.class, style=table.style, continue=table.continue} )
   else
-    CustomGameEventManager:Send_ServerToPlayer(player, "top_notification_heroimage", {hero=hero, imagestyle=imagestyle, duration=duration, class=class, style=style, continue=continue} )
+    CustomGameEventManager:Send_ServerToPlayer(player, "bottom_notification", {text="No TEXT provided.", duration=table.duration, class=table.class, style=table.style, continue=table.continue} )
   end
 end
 
-function Notifications:TopHeroImageToAll(hero, imagestyle, duration, class, style, continue)
-  if type(hero) == "table" then
-    CustomGameEventManager:Send_ServerToAllClients( "top_notification_heroimage", {hero=hero.hero, imagestyle=hero.imagestyle, duration=hero.duration, class=hero.class, style=hero.style, continue=hero.continue} )
+function Notifications:BottomToAll(table)
+  if table.text ~= nil then
+    CustomGameEventManager:Send_ServerToAllClients("bottom_notification", {text=table.text, duration=table.duration, class=table.class, style=table.style, continue=table.continue} )
+  elseif table.hero ~= nil then
+    CustomGameEventManager:Send_ServerToAllClients("bottom_notification", {hero=table.hero, imagestyle=table.imagestyle, duration=table.duration, class=table.class, style=table.style, continue=table.continue} )
+  elseif table.image ~= nil then
+    CustomGameEventManager:Send_ServerToAllClients("bottom_notification", {image=table.image, duration=table.duration, class=table.class, style=table.style, continue=table.continue} )
+  elseif table.ability ~= nil then
+    CustomGameEventManager:Send_ServerToAllClients("bottom_notification", {ability=table.ability, duration=table.duration, class=table.class, style=table.style, continue=table.continue} )
   else
-    CustomGameEventManager:Send_ServerToAllClients( "top_notification_heroimage", {hero=hero, imagestyle=imagestyle, duration=duration, class=class, style=style, continue=continue} )
+    CustomGameEventManager:Send_ServerToAllClients("bottom_notification", {text="No TEXT provided.", duration=table.duration, class=table.class, style=table.style, continue=table.continue} )
   end
 end
 
-function Notifications:TopHeroImageToTeam(team, hero, imagestyle,  duration, class, style, continue)
-  if type(hero) == "table" then
-    CustomGameEventManager:Send_ServerToTeam(team, "top_notification_heroimage", {hero=hero.hero, imagestyle=hero.imagestyle, duration=hero.duration, class=hero.class, style=hero.style, continue=hero.continue} )
+function Notifications:BottomToTeam(team, table)
+  if table.text ~= nil then
+    CustomGameEventManager:Send_ServerToTeam(team, "bottom_notification", {text=table.text, duration=table.duration, class=table.class, style=table.style, continue=table.continue} )
+  elseif table.hero ~= nil then
+    CustomGameEventManager:Send_ServerToTeam(team, "bottom_notification", {hero=table.hero, imagestyle=table.imagestyle, duration=table.duration, class=table.class, style=table.style, continue=table.continue} )
+  elseif table.image ~= nil then
+    CustomGameEventManager:Send_ServerToTeam(team, "bottom_notification", {image=table.image, duration=table.duration, class=table.class, style=table.style, continue=table.continue} )
+  elseif table.ability ~= nil then
+    CustomGameEventManager:Send_ServerToTeam(team, "bottom_notification", {ability=table.ability, duration=table.duration, class=table.class, style=table.style, continue=table.continue} )
   else
-    CustomGameEventManager:Send_ServerToTeam(team, "top_notification_heroimage", {hero=hero, imagestyle=imagestyle, duration=duration, class=class, style=style, continue=continue} )
-  end
-end
-
-
-
-
-
-function Notifications:Bottom(player, text, duration, class, style, continue)
-  if type(player) == "number" then
-    player = PlayerResource:GetPlayer(player)
-  end
-
-  if type(text) == "table" then
-    CustomGameEventManager:Send_ServerToPlayer(player, "bottom_notification", {text=text.text, duration=text.duration, class=text.class, style=text.style, continue=text.continue} )
-  else
-    CustomGameEventManager:Send_ServerToPlayer(player, "bottom_notification", {text=text, duration=duration, class=class, style=style, continue=continue} )
-  end
-end
-
-function Notifications:BottomToAll(text, duration, class, style, continue)
-  if type(text) == "table" then
-    CustomGameEventManager:Send_ServerToAllClients( "bottom_notification", {text=text.text, duration=text.duration, class=text.class, style=text.style, continue=text.continue} )
-  else
-    CustomGameEventManager:Send_ServerToAllClients( "bottom_notification", {text=text, duration=duration, class=class, style=style, continue=continue} )
-  end
-end
-
-function Notifications:BottomToTeam(team, text, duration, class, style, continue)
-  if type(text) == "table" then
-    CustomGameEventManager:Send_ServerToTeam(team, "bottom_notification", {text=text.text, duration=text.duration, class=text.class, style=text.style, continue=text.continue} )
-  else
-    CustomGameEventManager:Send_ServerToTeam(team, "bottom_notification", {text=text, duration=duration, class=class, style=style, continue=continue} )
-  end
-end
-
-
-function Notifications:BottomHeroImage(player, hero, imagestyle, duration, class, style, continue)
-  if type(player) == "number" then
-    player = PlayerResource:GetPlayer(player)
-  end
-
-  if type(hero) == "table" then
-    CustomGameEventManager:Send_ServerToPlayer(player, "bottom_notification_heroimage", {hero=hero.hero, imagestyle=hero.imagestyle, duration=hero.duration, class=hero.class, style=hero.style, continue=hero.continue} )
-  else
-    CustomGameEventManager:Send_ServerToPlayer(player, "bottom_notification_heroimage", {hero=hero, imagestyle=imagestyle, duration=duration, class=class, style=style, continue=continue} )
-  end
-end
-
-function Notifications:BottomHeroImageToAll(hero, imagestyle, duration, class, style)
-  if type(hero) == "table" then
-    CustomGameEventManager:Send_ServerToAllClients( "bottom_notification_heroimage", {hero=hero.hero, imagestyle=hero.imagestyle, duration=hero.duration, class=hero.class, style=hero.style, continue=hero.continue} )
-  else
-    CustomGameEventManager:Send_ServerToAllClients( "bottom_notification_heroimage", {hero=hero, imagestyle=imagestyle, duration=duration, class=class, style=style, continue=continue} )
-  end
-end
-
-function Notifications:BottomHeroImageToTeam(team, hero, imagestyle,  duration, class, style)
-  if type(hero) == "table" then
-    CustomGameEventManager:Send_ServerToTeam(team, "bottom_notification_heroimage", {hero=hero.hero, imagestyle=hero.imagestyle, duration=hero.duration, class=hero.class, style=hero.style, continue=hero.continue} )
-  else
-    CustomGameEventManager:Send_ServerToTeam(team, "bottom_notification_heroimage", {hero=hero, imagestyle=imagestyle, duration=duration, class=class, style=style, continue=continue} )
+    CustomGameEventManager:Send_ServerToTeam(team, "bottom_notification", {text="No TEXT provided.", duration=table.duration, class=table.class, style=table.style, continue=table.continue} )
   end
 end